Reestrutura codigo de prompt ao usuario
